@@ -1,10 +1,10 @@
 import test from 'ava';
-import m from './';
+import m from '.';
 
 test('send', async t => {
-	t.notThrows(m('green'));
+	await t.notThrows(m('green'));
 });
 
 test('custom port', async t => {
-	t.notThrows(m('red', {port: 1000}));
+	await t.notThrows(m('red', {port: 1000}));
 });
